+++ project/musics/views/sound_view.py
@@ -1,4 +1,5 @@
 from django.shortcuts import get_list_or_404, get_object_or_404
+from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -21,17 +22,35 @@ class SoundViewList(APIView):
     def post(self, request):
         print("SoundViewList.post")
         print("request.data", request.data)
+        if request.user.profile.is_coordinator is False:
+            return Response(
+                {"detail": "Só o coordenador por adicionar músicas"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
         soundSerializerUpsert = SoundSerializerUpsert(
-            data=request,
-            partial=True,
+            data=request.data,
         )
         soundSerializerUpsert.is_valid(
             raise_exception=True,
         )
+        audio = request.data.get("audio")
+        if bool(audio):
+            print("audio", audio)
+            audioNew = Audio.objects.create(
+                audio=audio,
+                profile=request.user.profile,
+            )
+        else:
+            return Response(
+                {"detail": "Não é possivel salvar uma música sem áudio."},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
         sound = Sound.objects.create(
-            name=soundSerializerUpsert.validated_data.get("name"),
-            description=soundSerializerUpsert.validated_data.get("description", None),
-            author=request.user.profile,
+            name=soundSerializerUpsert.validated_data.get("name", None),
+            description=request.data.get("description", None),
+            author=soundSerializerUpsert.validated_data.get("author", None),
+            album=soundSerializerUpsert.validated_data.get("album", None),
+            audio=audioNew,
         )
         image = request.data.get("image")
         if bool(image):
@@ -41,25 +60,20 @@ class SoundViewList(APIView):
                 profile=request.user.profile,
             )
             sound.image = imageNew
-        audio = request.data.get("audio")
-        if bool(audio):
-            print("audio", audio)
-            audioNew = Audio.objects.create(
-                audio=audio,
-                profile=request.user.profile,
-            )
-            sound.audio = audioNew
-
         sound.save()
 
-        Response({"id": sound.id})
+        return Response({"id": sound.id})
 
-    def get(self, request, id):
+    def get(self, request):
         print("SoundByAlbumViewList.get")
         print("request.data", request.data)
         albumId = request.GET.get("album", None)
         if albumId is None:
-            sounds = Sound.objects.all()
+            # sounds = Sound.objects.all()
+            return Response(
+                {"detail": "Só é possível listar musicas de algum album."},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
         else:
             sounds = get_list_or_404(
                 Sound.objects.all(),
@@ -69,7 +83,7 @@ class SoundViewList(APIView):
             sounds,
             many=True,
         )
-        Response(soundSerializerList.data)
+        return Response(soundSerializerList.data)
 
 
 class SoundViewDetail(APIView):
@@ -95,13 +109,17 @@ class SoundViewDetail(APIView):
         print("SoundViewDetail.post")
         print("request.data", request.data)
         print("id", id)
-
+        if request.user.profile.is_coordinator is False:
+            return Response(
+                {"detail": "Só o coordenador por editar músicas"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
         sound = get_object_or_404(
             Sound.objects.all(),
             id=id,
         )
         soundSerializerUpsert = SoundSerializerUpsert(
-            data=request,
+            data=request.data,
             partial=True,
         )
         soundSerializerUpsert.is_valid(
@@ -111,7 +129,7 @@ class SoundViewDetail(APIView):
             "name",
             sound.name,
         )
-        sound.description = soundSerializerUpsert.validated_data.get(
+        sound.description = request.data.get(
             "description",
             sound.description,
         )
@@ -155,9 +173,19 @@ class SoundViewDetail(APIView):
         print("SoundViewDetail.delete")
         print("request.data", request.data)
         print("id", id)
+        if request.user.profile.is_coordinator is False:
+            return Response(
+                {"detail": "Você não tem permissão para apagar músicas"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
         sound = get_object_or_404(
             Sound.objects.all(),
             id=id,
         )
+        if sound.album.coordinator is not request.user.profile:
+            return Response(
+                {"detail": "Você não tem permissão para apagar esta musica"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
         sound.delete()
         return Response()
